Add 4th arg
This change adds support for a 4th arg that passes a file for use as a launch splashscreen.
This is used with a modification to the runcommand.sh that allows for specifying a launch splashscreen as a 4th arg.
This has no impact when run with an unmodified runcommand.sh at this time. May cause issues with a future version that takes more args.